feat frontend: Created component for finish event

+++ frontend/src/app/(private)/calendar/details/[code]/useDatails.ts
@@ -2,7 +2,7 @@ import useApiPrivate from "@/hooks/apiPrivate";
 import { useQuery } from "@tanstack/react-query";
 import { queryClient } from "@/hooks/queryClient";
 import { ChangeEvent, useState } from "react";
-import { Clients } from "@/hooks/useClients";
+import { Event } from "@/types/events";
 
 const useDetails = (code: string) => {
   const api = useApiPrivate();
@@ -14,11 +14,12 @@ const useDetails = (code: string) => {
   const handleShowModalFinish = () => setShowModalFinish((prev) => !prev);
   const handleShowModalDelete = () => setShowModalDelete((prev) => !prev);
 
-  const { data: event, isLoading } = useQuery(["event", code], async () => {
+  const { data: event, isLoading } = useQuery(["event", code], async (): Promise<Event> => {
     return (await api.get(`/events/find/${code}`)).data;
   });
 
   async function deleteEvent() {
+    if(!event) return; 
     const res = await api.delete(`/events/delete/${event.id}`);
 
     if (!(res.statusText === "OK")) {
